fix 64bit types
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/branches/gnustep_testplant_branch@40150 72102866-910b-0410-8b05-ffd578937521

+++ Source/win32/NSMessagePort.m
@@ -288,7 +288,7 @@ static Class		messagePortClass = 0;
   [d handlePortMessage: m];
 }
 
-- (unsigned) hash
+- (NSUInteger) hash
 {
   return [PORT(self)->name hash];
 }
@@ -941,7 +941,7 @@ again:
  * the start can be written directly without having to copy data to another
  * buffer etc.
  */
-- (unsigned int) reservedSpaceLength
+- (NSUInteger) reservedSpaceLength
 {
   return sizeof(GSPortItemHeader) + sizeof(GSPortMsgHeader);
 }
